requirements.txt
 solana==0.9.0
setuptools~=54.2.0
uvicorn~=0.13.4
fastapi~=0.65.1
tinydb==4.4.0
requests==2.25.1
pydantic~=1.8.2
ipython==7.23.1
pyyaml==5.4.1
inflection==0.5.0
ip2geotools==0.1.5
supervisor
jinja2
aiofiles
Markdown==3.3.4
python-multipart==0.0.5
